# tacocloud/tacocloud-api/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Taco Cloud Public API
  version: 1.0.0
  description: |
    Contrato oficial OpenAPI 3.0 para Taco Cloud (Laboratorio 6 - TC-35).
    Define los contratos de integración para ingredientes, diseño de tacos, pedidos con idempotencia
    y trazabilidad por Correlation ID, favoritos, cupones, pasarela de pago tokenizada, operaciones de cocina y
    anuncios operativos seguros. Incluye el modelo canónico de error RFC 7807 (ApiProblem).
  contact:
    name: Taco Cloud Engineering Team
    email: support@tacocloud.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8080/api/v1
    description: Servidor de desarrollo local (v1 Oficial)
  - url: http://localhost:8080/api
    description: Servidor de desarrollo local (Alias legado - Deprecado)
...
paths:
...
  /ingredients/{id}:
    get:
      tags: [Ingredients]
      summary: Obtener un ingrediente por su identificador
      operationId: getIngredientById
      parameters:
        - name: id
          in: path
          required: true
          description: Código del ingrediente (ej. FLTO, CARN)
          schema:
            type: string
            example: FLTO
      responses:
        '200':
          description: Ingrediente encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngredientResponse'
        '404':
          description: Ingrediente no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
    delete:
      tags: [Ingredients]
      summary: Eliminar un ingrediente (Admin)
      operationId: deleteIngredient
      security:
        - basicAuth: []
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Ingrediente eliminado exitosamente
        '403':
          description: Requiere rol ADMIN
...
  /orders/{orderId}:
    get:
      tags: [Orders]
      summary: Consultar un pedido por su identificador
      operationId: getOrderById
      security:
        - basicAuth: []
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Detalle del pedido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '404':
          description: Pedido no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
    delete:
      tags: [Orders]
      summary: Cancelar un pedido
      operationId: cancelOrder
      security:
        - basicAuth: []
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Pedido cancelado
...
  /users/me/orders/{id}:
    get:
      tags: [User Profile]
      summary: Consultar detalle de un pedido propio
      operationId: getMyOrderDetail
      security:
        - basicAuth: []
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Detalle del pedido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDetailResponse'
        '404':
          description: Pedido no encontrado o no pertenece al usuario
...
  /coupons/validate:
    post:
      tags: [Coupons]
      summary: Validar código promocional y calcular descuento
      operationId: validateCoupon
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CouponValidateRequest'
      responses:
        '200':
          description: Resultado de la validación del cupón
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CouponValidateResponse'
...
  /kitchen/queue:
    get:
      tags: [Kitchen]
      summary: Consultar cola de pedidos pendientes de cocina (FIFO)
      operationId: getKitchenQueue
      security:
        - basicAuth: []
        - bearerAuth: []
      responses:
        '200':
          description: Cola de pedidos en espera
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/KitchenOrderDto'
  /kitchen/orders/claim:
    post:
      tags: [Kitchen]
      summary: Reclamar atómicamente la siguiente orden para una estación
      operationId: claimKitchenOrder
      security:
        - basicAuth: []
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KitchenClaimRequest'
      responses:
        '200':
          description: Orden reclamada para preparación
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KitchenOrderDto'
        '204':
          description: No hay órdenes disponibles en la cola
  /kitchen/orders/{id}/status:
    patch:
      tags: [Kitchen]
      summary: Actualizar estado de preparación de un pedido en cocina
      operationId: updateKitchenOrderStatus
      security:
        - basicAuth: []
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KitchenStatusUpdateRequest'
      responses:
        '200':
          description: Estado actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KitchenOrderDto'
...
  /admin/announcements:
    get:
      tags: [Announcements]
      summary: Listar todos los anuncios operativos (Admin)
      operationId: getAllAnnouncementsAdmin
      security:
        - basicAuth: []
        - bearerAuth: []
      responses:
        '200':
          description: Todos los anuncios (activos y expirados)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OpsAnnouncementAdminDto'
    post:
      tags: [Announcements]
      summary: Publicar un nuevo anuncio operativo (Admin)
      operationId: createAnnouncement
      security:
        - basicAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpsAnnouncementRequest'
      responses:
        '201':
          description: Anuncio publicado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpsAnnouncementAdminDto'
        '400':
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT para usuarios autenticados
    basicAuth:
      type: http
      scheme: basic
      description: Autenticación básica HTTP (usuario y contraseña)
...
  schemas:
...
    IngredientResponse:
      type: object
      required:
        - id
        - name
        - type
      properties:
        id:
          type: string
          example: FLTO
        name:
          type: string
          example: Flour Tortilla
        type:
          type: string
          example: WRAP

    # --- TACOS ---
...
    TacoResponse:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
          example: 64b8a1c9e4b0c23a8f123456
        name:
          type: string
          example: Carnitas Fiesta
        createdAt:
          type: string
          format: date-time
        ingredients:
          type: array
          items:
            $ref: '#/components/schemas/IngredientResponse'
        dietaryTags:
          type: array
          items:
            type: string
          example: [GLUTEN_FREE]
        allergens:
          type: array
          items:
            type: string
          example: [DAIRY]
        spiceLevel:
          type: string
          example: MEDIUM
...
    OrderItemResponse:
      type: object
      properties:
        taco:
          $ref: '#/components/schemas/TacoResponse'
        quantity:
          type: integer
          example: 2
        unitPrice:
          type: number
          format: double
          example: 4.50
        lineTotal:
          type: number
          format: double
          example: 9.00
...
    OrderResponse:
      type: object
      required:
        - id
        - deliveryName
        - total
      properties:
        id:
          type: string
          example: 64b8a1c9e4b0c23a8f999999
        placedAt:
          type: string
          format: date-time
        deliveryName:
          type: string
          example: Jane Doe
        deliveryStreet:
          type: string
          example: 123 Main St
        deliveryCity:
          type: string
          example: Austin
        deliveryState:
          type: string
          example: TX
        deliveryZip:
          type: string
          example: 78701
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemResponse'
        subtotal:
          type: number
          format: double
          example: 18.00
        discountAmount:
          type: number
          format: double
          example: 3.60
        total:
          type: number
          format: double
          example: 14.40
        currency:
          type: string
          example: USD
        status:
          type: string
          example: PLACED
        brand:
          type: string
          example: Visa
        last4:
          type: string
          example: '4242'
...
    OrderDetailResponse:
      type: object
      properties:
        id:
          type: string
          example: 64b8a1c9e4b0c23a8f999999
        placedAt:
          type: string
          format: date-time
        deliveryName:
          type: string
          example: Jane Doe
        deliveryCity:
          type: string
          example: Austin
        total:
          type: number
          example: 14.40
        status:
          type: string
          example: READY

    # --- FAVORITOS ---
...
    CouponValidateRequest:
      type: object
      required:
        - code
        - subtotal
      properties:
        code:
          type: string
          example: TACO20
        subtotal:
          type: number
          format: double
          example: 25.00
    CouponValidateResponse:
      type: object
      properties:
        valid:
          type: boolean
          example: true
        code:
          type: string
          example: TACO20
        status:
          type: string
          example: APPLIED
        message:
          type: string
          example: 20% discount applied
        discountAmount:
          type: number
          format: double
          example: 5.00
        finalTotal:
          type: number
          format: double
          example: 20.00
        currency:
          type: string
          example: USD

    # --- PAGOS (SIN CAMPOS SENSIBLES) ---
...
    KitchenOrderDto:
      type: object
      properties:
        orderId:
          type: string
          example: 64b8a1c9e4b0c23a8f999999
        status:
          type: string
          example: PREPARING
        station:
          type: string
          example: STATION-1
        placedAt:
          type: string
          format: date-time
    KitchenClaimRequest:
      type: object
      properties:
        station:
          type: string
          example: STATION-1
    KitchenStatusUpdateRequest:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum: [PREPARING, READY, DELIVERED]
          example: READY
...
    OpsAnnouncementRequest:
      type: object
      required:
        - message
        - severity
      properties:
        message:
          type: string
          example: Cocina operando con demoras por alta demanda
        severity:
          type: string
          enum: [INFO, WARNING, CRITICAL]
          example: WARNING
        expiresInMinutes:
          type: integer
          example: 60
...
    OpsAnnouncementAdminDto:
      type: object
      properties:
        id:
          type: string
          example: ann-1234
        message:
          type: string
          example: Cocina operando con demoras por alta demanda
        severity:
          type: string
          example: WARNING
        createdBy:
          type: string
          example: ops_admin
        active:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time

    # --- MODELO ESTÁNDAR DE ERROR (RFC 7807) ---
    ApiProblem:
      type: object
      required:
        - title
        - status
      properties:
        type:
          type: string
          format: uri
          example: about:blank
        title:
          type: string
          example: Bad Request
        status:
          type: integer
          example: 400
        detail:
          type: string
          example: Request validation failed
        instance:
          type: string
          example: /api/v1/orders
        code:
          type: string
          example: VALIDATION_ERROR
        violations:
          type: array
          items:
            $ref: '#/components/schemas/ApiProblemViolation'
        timestamp:
          type: string
          format: date-time
        traceId:
          type: string
          example: 550e8400-e29b-41d4-a716-446655440000
    ApiProblemViolation:
      type: object
      properties:
        field:
          type: string
          example: deliveryName
        message:
          type: string
          example: Delivery name is required